Finalized part 2

+++ src/edu/tamu/csce315_908_t4/imdbParser/outputDataType/Rating.java
@@ -3,10 +3,10 @@ package edu.tamu.csce315_908_t4.imdbParser.outputDataType;
 @SuppressWarnings("WeakerAccess")
 public class Rating implements IOutputTable{
     public final String tconst;
-    public final float averageRating;
-    public final long numVotes;
+    public final Float averageRating;
+    public final Integer numVotes;
 
-    public Rating(String tconst, float averageRating, long numVotes){
+    public Rating(String tconst, Float averageRating, Integer numVotes){
         this.tconst = tconst;
         this.averageRating = averageRating;
         this.numVotes = numVotes;
